Unminimize image before installing all packages.
I couldn't find much data on the topic online, yet inspecting the
image's build logs led me to conclude that the unminimizing process causes many of the
packages to reinstall. Thus it may be preferable to firstly unminimize.

KEEP IN MIND: I decided to move the unminimize command to a seperate
layer. This may potentially lead to cache errors between the layers.